+++ src/Algorithms/Sort_methods/Merge_sort.h
@@ -17,7 +17,7 @@
 
 namespace CZ
 {
-    using Rank_merge_sort = long long;
+    using Rank_merge_sort = int64_t;
 
     namespace MergeSort
     {
