+++ src/Algorithms/Sort_methods/Shell_sort.h
@@ -18,7 +18,7 @@
 
 namespace CZ
 {
-    using Rank_shell_sort = long long;
+    using Rank_shell_sort = int64_t;
 
     namespace ShellSort
     {
