+++ src/Dictionary/HashTable.h
@@ -22,7 +22,7 @@ namespace CZ
 {
     namespace Dict
     {
-        using Rank = int;
+        using Rank = int64_t;
 
         // 哈希函数，取模法
         template <typename T>
